@@ -17,65 +17,63 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.0.1",
-    "@angular/common": "^17.0.1",
-    "@angular/compiler": "^17.0.1",
-    "@angular/core": "^17.0.1",
-    "@angular/forms": "^17.3.0",
-    "@angular/localize": "^17.3.0",
-    "@angular/platform-browser": "^17.0.1",
-    "@angular/platform-browser-dynamic": "^17.3.0",
-    "@angular/router": "^17.3.0",
-    "@ngxs/router-plugin": "^3.8.2",
-    "@ngxs/storage-plugin": "^3.8.2",
-    "@ngxs/store": "^3.8.2",
+    "@angular/animations": "^18.0.1",
+    "@angular/common": "^18.0.1",
+    "@angular/compiler": "^18.0.1",
+    "@angular/core": "^18.0.1",
+    "@angular/forms": "^18.0.3",
+    "@angular/localize": "^18.0.3",
+    "@angular/platform-browser": "^18.0.1",
+    "@angular/platform-browser-dynamic": "^18.0.3",
+    "@angular/router": "^18.0.3",
+    "@ngxs/router-plugin": "^18.0.0",
+    "@ngxs/storage-plugin": "^18.0.0",
+    "@ngxs/store": "^18.0.0",
     "crypto-js": "^4.2.0",
     "fast-json-patch": "^3.1.1",
     "moment": "^2.30.1",
     "primeflex": "^3.3.1",
-    "primeicons": "^6.0.1",
-    "primeng": "^17.11.0",
+    "primeicons": "^7.0.0",
+    "primeng": "^17.18.1",
     "rxjs": "^7.8.1",
-    "tslib": "^2.6.2",
-    "zone.js": "^0.14.4"
+    "tslib": "^2.6.3",
+    "zone.js": "^0.14.7"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.0",
-    "@angular-eslint/builder": "^17.2.1",
-    "@angular-eslint/eslint-plugin": "^17.0.0",
-    "@angular-eslint/eslint-plugin-template": "^17.0.0",
-    "@angular-eslint/schematics": "^17.2.1",
-    "@angular-eslint/template-parser": "^17.2.1",
-    "@angular/cli": "^17.3.0",
-    "@angular/compiler-cli": "^17.0.1",
-    "@angular/language-service": "^17.3.0",
-    "@badeball/cypress-cucumber-preprocessor": "^20.0.2",
-    "@cypress/schematic": "^2.5.1",
-    "@cypress/webpack-preprocessor": "^6.0.1",
+    "@angular-devkit/build-angular": "^18.0.4",
+    "@angular-eslint/builder": "^18.0.1",
+    "@angular-eslint/eslint-plugin": "^18.0.1",
+    "@angular-eslint/eslint-plugin-template": "^18.0.1",
+    "@angular-eslint/schematics": "^18.0.1",
+    "@angular-eslint/template-parser": "^18.0.1",
+    "@angular/cli": "^18.0.4",
+    "@angular/compiler-cli": "^18.0.1",
+    "@angular/language-service": "^18.0.3",
+    "@badeball/cypress-cucumber-preprocessor": "^20.0.7",
+    "@cypress/schematic": "^2.5.2",
+    "@cypress/webpack-preprocessor": "^6.0.2",
     "@sinonjs/fake-timers": "^11.2.2",
-    "@types/chai": "^4.3.12",
+    "@types/chai": "^4.3.16",
     "@types/mocha": "^10.0.6",
-    "@types/node": "^20.11.27",
+    "@types/node": "^20.14.2",
     "@types/sinon": "^17.0.3",
-    "@typescript-eslint/eslint-plugin": "^7.2.0",
-    "@typescript-eslint/parser": "^7.2.0",
+    "@typescript-eslint/eslint-plugin": "^7.13.0",
+    "@typescript-eslint/parser": "^7.13.0",
     "cassandra-driver": "^4.7.2",
-    "chai": "^5.1.0",
-    "cypress": "^13.7.0",
+    "chai": "^5.1.1",
+    "cypress": "^13.11.0",
     "eslint": "^8.57.0",
     "karma": "^6.4.3",
     "karma-chrome-launcher": "^3.2.0",
     "karma-coverage": "^2.2.1",
     "karma-mocha": "^2.0.1",
     "karma-mocha-reporter": "^2.2.5",
-    "mocha": "^10.3.0",
-    "ng-mocks": "^14.12.2",
-    "process": "^0.11.10",
-    "sinon": "^17.0.1",
-    "sonarqube-scanner": "^3.3.0",
+    "mocha": "^10.4.0",
+    "ng-mocks": "^14.13.0",
+    "sinon": "^18.0.0",
+    "sonarqube-scanner": "^4.0.1",
     "ts-loader": "^9.5.1",
-    "typescript": "^5.4.2",
-    "util": "^0.12.5"
+    "typescript": "^5.4.5"
   },
   "cypress-cucumber-preprocessor": {
     "nonGlobalStepDefinitions": true,
