@@ -60,9 +60,7 @@ export class AccountEditComponent implements OnInit {
   save() {
 
     const account = { ...this.accountForm.value };
-    this.store.dispatch(new UpdateAccount(account, this.account)).subscribe(() =>
-      this.account = this.accountForm.value
-    );
+    this.store.dispatch(new UpdateAccount(account, this.account));
   }
 
   cancel() {
