@@ -1,6 +1,7 @@
 import { DebugElement } from '@angular/core';
 import { inject, TestBed, ComponentFixture, waitForAsync } from '@angular/core/testing';
-import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
+import { HttpTestingController, provideHttpClientTesting } from '@angular/common/http/testing';
 import { By } from '@angular/platform-browser';
 import { expect } from 'chai';
 
@@ -8,51 +9,50 @@ import { InfoComponent } from './info.component';
 
 describe('InfoComponent', () => {
 
-    let fixture: ComponentFixture<InfoComponent>;
-    let debugElement: DebugElement;
+  let fixture: ComponentFixture<InfoComponent>;
+  let debugElement: DebugElement;
 
-    beforeEach(waitForAsync(() => {
+  beforeEach(waitForAsync(() => {
 
-        fixture = TestBed.configureTestingModule({
+    fixture = TestBed.configureTestingModule({
+      imports: [],
+      providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+    }).createComponent(InfoComponent);
 
-            imports: [HttpClientTestingModule]
 
-        }).createComponent(InfoComponent);
+  }));
 
+  beforeEach(() => {
 
-    }));
+    fixture.detectChanges();
+    debugElement = fixture.debugElement;
 
-    beforeEach(() => {
+  });
 
-        fixture.detectChanges();
-        debugElement = fixture.debugElement;
+  afterEach(() => {
 
-    });
+    TestBed.resetTestingModule();
 
-    afterEach(() => {
+  });
 
-        TestBed.resetTestingModule();
+  it('display info', waitForAsync(inject(
+    [HttpTestingController], (http) => {
 
-    });
+      const req = http.expectOne('/ExempleService/actuator/info');
+      req.flush({
+        version: 'test',
+        buildTime: '1976-01-03'
+      });
 
-    it('display info', waitForAsync(inject(
-        [HttpTestingController], (http) => {
+      http.verify();
 
-            const req = http.expectOne('/ExempleService/actuator/info');
-            req.flush({
-                version: 'test',
-                buildTime: '1976-01-03'
-            });
+      fixture.detectChanges();
 
-            http.verify();
+      const de: DebugElement[] = debugElement.queryAll(By.css('h4'));
 
-            fixture.detectChanges();
+      expect(de[0].nativeElement.innerHTML).to.equal('version:test');
+      expect(de[1].nativeElement.innerHTML).to.equal('buildTime:1976-01-03');
 
-            const de: DebugElement[] = debugElement.queryAll(By.css('h4'));
-
-            expect(de[0].nativeElement.innerHTML).to.equal('version:test');
-            expect(de[1].nativeElement.innerHTML).to.equal('buildTime:1976-01-03');
-
-        })));
+    })));
 
 });
