@@ -1,17 +1,19 @@
 import { Injectable } from '@angular/core';
-import { Action, State, StateContext } from '@ngxs/store';
+import { Action, State, StateContext, StateToken } from '@ngxs/store';
 import { Message } from 'primeng/api';
 
 import { PublishMessage } from './message.action';
 
-@State<Message[]>({
-    name: 'messages'
+export const MESSAGE_STATE_TOKEN = new StateToken<Message>('message');
+
+@State<Message>({
+  name: MESSAGE_STATE_TOKEN
 })
 @Injectable()
 export class MessageState {
 
-    @Action(PublishMessage)
-    publish(ctx: StateContext<Message>, action: PublishMessage) {
-        ctx.patchState(action.message);
-    }
+  @Action(PublishMessage)
+  publish(ctx: StateContext<Message>, action: PublishMessage) {
+    ctx.patchState(action.message);
+  }
 }
